Move all thunk support code into thunks.cpp and thunks.h
